+++ infrastructures/symfony/Router/Router.php
@@ -30,7 +30,7 @@ use ReflectionClass;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController as SfAbstractController;
 use Symfony\Component\DependencyInjection\ContainerInterface;
 use Symfony\Component\Routing\Exception\ResourceNotFoundException;
-use Teknoo\East\Foundation\Http\ClientInterface;
+use Teknoo\East\Foundation\Client\ClientInterface;
 use Teknoo\East\Foundation\Manager\ManagerInterface;
 use Teknoo\East\Foundation\Middleware\MiddlewareInterface;
 use Psr\Http\Message\ServerRequestInterface;
@@ -120,7 +120,7 @@ class Router implements RouterInterface
                     return null;
                 }
 
-                //Method must exist because the controller is callable
+                //Method must exist because the controller is a symfony action callable
                 $method = $reflection->getMethod($explodedController[1]);
                 if (!$method->isStatic()) {
                     return null;