+++ src/Middleware/MiddlewareInterface.php
@@ -26,7 +26,7 @@ declare(strict_types=1);
 namespace Teknoo\East\Foundation\Middleware;
 
 use Psr\Http\Message\MessageInterface;
-use Teknoo\East\Foundation\Http\ClientInterface;
+use Teknoo\East\Foundation\Client\ClientInterface;
 use Teknoo\East\Foundation\Manager\ManagerInterface;
 
 /**