+++ tests/infrastructures/symfony/Http/ClientTest.php
@@ -23,12 +23,15 @@
 namespace Teknoo\Tests\East\FoundationBundle\Http;
 
 use PHPUnit\Framework\TestCase;
+use Psr\Http\Message\ResponseFactoryInterface;
 use Psr\Http\Message\ResponseInterface;
+use Psr\Http\Message\StreamFactoryInterface;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Bridge\PsrHttpMessage\Factory\HttpFoundationFactory;
 use Symfony\Component\HttpKernel\Event\RequestEvent;
-use Teknoo\East\Foundation\Http\ClientInterface;
+use Teknoo\East\Foundation\Client\ClientInterface;
+use Teknoo\East\Foundation\Client\ResponseInterface as EastResponse;
 use Teknoo\East\FoundationBundle\Http\Client;
 
 /**
@@ -55,6 +58,10 @@ class ClientTest extends TestCase
      */
     private $httpFoundationFactory;
 
+    private ?ResponseFactoryInterface $responseFactory = null;
+
+    private ?StreamFactoryInterface $streamFactory = null;
+
     /**
      * @return RequestEvent|\PHPUnit\Framework\MockObject\MockObject
      */
@@ -80,12 +87,45 @@ class ClientTest extends TestCase
     }
 
     /**
+     * @return ResponseFactoryInterface|\PHPUnit\Framework\MockObject\MockObject
+     */
+    private function getResponseFactoryMock(): ResponseFactoryInterface
+    {
+        if (!$this->responseFactory instanceof ResponseFactoryInterface) {
+            $this->responseFactory = $this->createMock(ResponseFactoryInterface::class);
+
+            $response = $this->createMock(ResponseInterface::class);
+            $response->expects(self::any())->method('withBody')->willReturnSelf();
+
+            $this->responseFactory->expects(self::any())
+                ->method('createResponse')
+                ->willReturn($response);
+        }
+
+        return $this->responseFactory;
+    }
+
+    /**
+     * @return StreamFactoryInterface|\PHPUnit\Framework\MockObject\MockObject
+     */
+    private function getStreamFactoryMock(): StreamFactoryInterface
+    {
+        if (!$this->streamFactory instanceof StreamFactoryInterface) {
+            $this->streamFactory = $this->createMock(StreamFactoryInterface::class);
+        }
+
+        return $this->streamFactory;
+    }
+
+    /**
      * @return Client
      */
     private function buildClient(LoggerInterface $logger = null): Client
     {
         return new Client(
             $this->getHttpFoundationFactoryMock(),
+            $this->getResponseFactoryMock(),
+            $this->getStreamFactoryMock(),
             $this->getRequestEventMock(),
             $logger
         );
@@ -116,7 +156,7 @@ class ClientTest extends TestCase
         );
     }
 
-    public function testUpdateResponseWithResponse()
+    public function testUpdateResponseWithPSRResponse()
     {
         /**
          * @var ResponseInterface
@@ -134,13 +174,57 @@ class ClientTest extends TestCase
         );
     }
 
+    public function testUpdateResponseWithEastResponse()
+    {
+        /**
+         * @var ResponseInterface
+         */
+        $response = $this->createMock(EastResponse::class);
+
+        $client = $this->buildClient();
+        self::assertInstanceOf(
+            $this->getClientClass(),
+            $client->acceptResponse($response)->updateResponse(
+                function (ClientInterface $client, EastResponse $responsePassed=null) use ($response) {
+                    self::assertEquals($response, $responsePassed);
+                }
+            )
+        );
+    }
+
+    public function testUpdateResponseWithJsonResponse()
+    {
+        $response = new class implements EastResponse, \JsonSerializable
+        {
+            public function __toString(): string
+            {
+                return 'foo';
+            }
+
+            public function jsonSerialize()
+            {
+                return ['foo' => 'bar'];
+            }
+        };
+
+        $client = $this->buildClient();
+        self::assertInstanceOf(
+            $this->getClientClass(),
+            $client->acceptResponse($response)->updateResponse(
+                function (ClientInterface $client, EastResponse $responsePassed=null) use ($response) {
+                    self::assertEquals($response, $responsePassed);
+                }
+            )
+        );
+    }
+
     public function testAcceptResponseError()
     {
         $this->expectException(\TypeError::class);
         $this->buildClient()->acceptResponse(new \stdClass());
     }
     
-    public function testAcceptResponse()
+    public function testAcceptPSRResponse()
     {
         /**
          * @var ResponseInterface
@@ -153,8 +237,41 @@ class ClientTest extends TestCase
             $client->acceptResponse($response)
         );
     }
+
+    public function testAcceptEastResponse()
+    {
+        $response = $this->createMock(EastResponse::class);
+
+        $client = $this->buildClient();
+        self::assertInstanceOf(
+            $this->getClientClass(),
+            $client->acceptResponse($response)
+        );
+    }
+
+    public function testAcceptJsonResponse()
+    {
+        $response = new class implements EastResponse, \JsonSerializable
+        {
+            public function __toString(): string
+            {
+                return 'foo';
+            }
+
+            public function jsonSerialize()
+            {
+                return ['foo' => 'bar'];
+            }
+        };
+
+        $client = $this->buildClient();
+        self::assertInstanceOf(
+            $this->getClientClass(),
+            $client->acceptResponse($response)
+        );
+    }
     
-    public function testSendResponse()
+    public function testSendPSRResponse()
     {
         /**
          * @var ResponseInterface
@@ -175,7 +292,7 @@ class ClientTest extends TestCase
             ->with($this->callback(function ($response) {
                 return $response instanceof ResponseInterface;
             }))
-            ->willReturn($this->createMock(Response::class, [], [], '', false));
+            ->willReturn($this->createMock(Response::class));
 
         $client = $this->buildClient();
         self::assertInstanceOf(
@@ -184,7 +301,72 @@ class ClientTest extends TestCase
         );
     }
 
-    public function testSendResponseWithAccept()
+    public function testSendEastResponse()
+    {
+        $response = $this->createMock(EastResponse::class);
+
+        $this->getRequestEventMock()
+            ->expects(self::any())
+            ->method('setResponse')
+            ->with($this->callback(function ($response) {
+                return $response instanceof Response;
+            }))
+            ->willReturnSelf();
+
+        $this->getHttpFoundationFactoryMock()
+            ->expects(self::any())
+            ->method('createResponse')
+            ->with($this->callback(function ($response) {
+                return $response instanceof ResponseInterface;
+            }))
+            ->willReturn($this->createMock(Response::class));
+
+        $client = $this->buildClient();
+        self::assertInstanceOf(
+            $this->getClientClass(),
+            $client->sendResponse($response)
+        );
+    }
+
+    public function testSendJsonResponse()
+    {
+        $response = new class implements EastResponse, \JsonSerializable
+        {
+            public function __toString(): string
+            {
+                return 'foo';
+            }
+
+            public function jsonSerialize()
+            {
+                return ['foo' => 'bar'];
+            }
+        };
+
+        $this->getRequestEventMock()
+            ->expects(self::any())
+            ->method('setResponse')
+            ->with($this->callback(function ($response) {
+                return $response instanceof Response;
+            }))
+            ->willReturnSelf();
+
+        $this->getHttpFoundationFactoryMock()
+            ->expects(self::any())
+            ->method('createResponse')
+            ->with($this->callback(function ($response) {
+                return $response instanceof ResponseInterface;
+            }))
+            ->willReturn($this->createMock(Response::class));
+
+        $client = $this->buildClient();
+        self::assertInstanceOf(
+            $this->getClientClass(),
+            $client->sendResponse($response)
+        );
+    }
+
+    public function testSendPSRResponseWithAccept()
     {
         /**
          * @var ResponseInterface
@@ -205,7 +387,72 @@ class ClientTest extends TestCase
             ->with($this->callback(function ($response) {
                 return $response instanceof ResponseInterface;
             }))
-            ->willReturn($this->createMock(Response::class, [], [], '', false));
+            ->willReturn($this->createMock(Response::class));
+
+        $client = $this->buildClient();
+        self::assertInstanceOf(
+            $this->getClientClass(),
+            $client->acceptResponse($response)->sendResponse()
+        );
+    }
+
+    public function testSendEastResponseWithAccept()
+    {
+        $response = $this->createMock(EastResponse::class);
+
+        $this->getRequestEventMock()
+            ->expects(self::any())
+            ->method('setResponse')
+            ->with($this->callback(function ($response) {
+                return $response instanceof Response;
+            }))
+            ->willReturnSelf();
+
+        $this->getHttpFoundationFactoryMock()
+            ->expects(self::any())
+            ->method('createResponse')
+            ->with($this->callback(function ($response) {
+                return $response instanceof ResponseInterface;
+            }))
+            ->willReturn($this->createMock(Response::class));
+
+        $client = $this->buildClient();
+        self::assertInstanceOf(
+            $this->getClientClass(),
+            $client->acceptResponse($response)->sendResponse()
+        );
+    }
+
+    public function testSendJsonResponseWithAccept()
+    {
+        $response = new class implements EastResponse, \JsonSerializable
+        {
+            public function __toString(): string
+            {
+                return 'foo';
+            }
+
+            public function jsonSerialize()
+            {
+                return ['foo' => 'bar'];
+            }
+        };
+
+        $this->getRequestEventMock()
+            ->expects(self::any())
+            ->method('setResponse')
+            ->with($this->callback(function ($response) {
+                return $response instanceof Response;
+            }))
+            ->willReturnSelf();
+
+        $this->getHttpFoundationFactoryMock()
+            ->expects(self::any())
+            ->method('createResponse')
+            ->with($this->callback(function ($response) {
+                return $response instanceof ResponseInterface;
+            }))
+            ->willReturn($this->createMock(Response::class));
 
         $client = $this->buildClient();
         self::assertInstanceOf(
@@ -233,7 +480,12 @@ class ClientTest extends TestCase
          */
         $response = $this->createMock(ResponseInterface::class);
 
-        $client = new Client($this->getHttpFoundationFactoryMock());
+        $client = new Client(
+            $this->getHttpFoundationFactoryMock(),
+            $this->getResponseFactoryMock(),
+            $this->getStreamFactoryMock()
+        );
+
         self::assertInstanceOf(
             $this->getClientClass(),
             $client->sendResponse($response)
@@ -261,7 +513,7 @@ class ClientTest extends TestCase
             ->with($this->callback(function ($response) {
                 return $response instanceof ResponseInterface;
             }))
-            ->willReturn($this->createMock(Response::class, [], [], '', false));
+            ->willReturn($this->createMock(Response::class));
 
         $client = $this->buildClient();
         self::assertInstanceOf(
@@ -291,7 +543,7 @@ class ClientTest extends TestCase
             ->with($this->callback(function ($response) {
                 return $response instanceof ResponseInterface;
             }))
-            ->willReturn($this->createMock(Response::class, [], [], '', false));
+            ->willReturn($this->createMock(Response::class));
 
         $client = $this->buildClient();
         self::assertInstanceOf(
@@ -321,7 +573,7 @@ class ClientTest extends TestCase
             ->with($this->callback(function ($response) {
                 return $response instanceof ResponseInterface;
             }))
-            ->willReturn($this->createMock(Response::class, [], [], '', false));
+            ->willReturn($this->createMock(Response::class));
 
         $client = $this->buildClient();
         self::assertInstanceOf(
@@ -347,7 +599,11 @@ class ClientTest extends TestCase
          */
         $response = $this->createMock(ResponseInterface::class);
 
-        $client = new Client($this->getHttpFoundationFactoryMock());
+        $client = new Client(
+            $this->getHttpFoundationFactoryMock(),
+            $this->getResponseFactoryMock(),
+            $this->getStreamFactoryMock()
+        );
         self::assertInstanceOf(
             $this->getClientClass(),
             $client->sendResponse($response, true)
@@ -415,7 +671,11 @@ class ClientTest extends TestCase
     public function testErrorInRequestWithoutRequestEvent()
     {
         $this->expectException(\Exception::class);
-        $client = new Client($this->getHttpFoundationFactoryMock());
+        $client = new Client(
+            $this->getHttpFoundationFactoryMock(),
+            $this->getResponseFactoryMock(),
+            $this->getStreamFactoryMock()
+        );
         self::assertInstanceOf(
             $this->getClientClass(),
             $client->errorInRequest(new \Exception('fooBar'))